+++ src/core/ddsi/include/dds/ddsi/ddsi_ssl.h
@@ -24,7 +24,19 @@
 extern "C" {
 #endif
 
-struct ddsi_ssl_plugins;
+struct ddsi_ssl_plugins
+{
+  bool (*init) (struct ddsi_domaingv *gv);
+  void (*fini) (void);
+  void (*ssl_free) (SSL *ssl);
+  void (*bio_vfree) (BIO *bio);
+  ssize_t (*read) (SSL *ssl, void *buf, size_t len, dds_return_t *err);
+  ssize_t (*write) (SSL *ssl, const void *msg, size_t len, dds_return_t *err);
+  SSL * (*connect) (const struct ddsi_domaingv *gv, ddsrt_socket_t sock);
+  BIO * (*listen) (ddsrt_socket_t sock);
+  SSL * (*accept) (const struct ddsi_domaingv *gv, BIO *bio, ddsrt_socket_t *sock);
+};
+
 void ddsi_ssl_config_plugin (struct ddsi_ssl_plugins *plugin);
 
 #if defined (__cplusplus)
